Used border bottom for underline and renamed UserProfile.js->index.js
