+++ meow/frontend/src/components/UserProfile/UserProfileBasicInfo.js
@@ -8,34 +8,22 @@ function UserProfileBasicInfo(props) {
       <ul className="user-profile-basic-info-table">
         <li className="user-profile-basic-info-form-group">
           <p className="user-profile-basic-info-title">name: </p>
-          <div className="user-profile-basic-info-data">
-            {props.name}
-            <hr className="user-profile-basic-info-horizontal-line" />
-          </div>
+          <div className="user-profile-basic-info-data">{props.name}</div>
         </li>
 
         <li className="user-profile-basic-info-form-group">
           <p className="user-profile-basic-info-title">role: </p>
-          <div className="user-profile-basic-info-data">
-            {props.role}
-            <hr className="user-profile-basic-info-horizontal-line" />
-          </div>
+          <div className="user-profile-basic-info-data">{props.role}</div>
         </li>
 
         <li className="user-profile-basic-info-form-group">
           <p className="user-profile-basic-info-title">email: </p>
-          <div className="user-profile-basic-info-data">
-            {props.email}
-            <hr className="user-profile-basic-info-horizontal-line" />
-          </div>
+          <div className="user-profile-basic-info-data">{props.email}</div>
         </li>
 
         <li className="user-profile-basic-info-form-group">
           <p className="user-profile-basic-info-title">slack: </p>
-          <div className="user-profile-basic-info-data">
-            {props.slack_username}
-            <hr className="user-profile-basic-info-horizontal-line" />
-          </div>
+          <div className="user-profile-basic-info-data">{props.slack_username}</div>
         </li>
 
         <li className="user-profile-basic-info-form-group">
