+++ meow/frontend/src/components/UserProfile/styling.css
@@ -56,6 +56,7 @@
   /* max-width: 220px; */
   font-size: 36px;
   margin-bottom: 0px;
+  border-bottom: 2px solid #000;
 }
 
 .user-profile-basic-info-horizontal-line {
